feat: XYNE-93 added zustand for state managment
