+++ frontend/src/main.tsx
@@ -7,7 +7,6 @@ import { RouterProvider, createRouter } from "@tanstack/react-router"
 import { routeTree } from "@/routeTree.gen"
 import { ThemeProvider } from "@/components/ThemeContext"
 import { Toaster } from "@/components/ui/toaster"
-import { UploadProgressProvider } from "@/contexts/UploadProgressContext"
 import UploadProgressWidget from "@/components/UploadProgressWidget"
 
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
@@ -29,11 +28,9 @@ const App = () => {
   return (
     <ThemeProvider>
       <QueryClientProvider client={queryClient}>
-        <UploadProgressProvider>
-          <RouterProvider router={router} />
-          <Toaster />
-          <UploadProgressWidget />
-        </UploadProgressProvider>
+        <RouterProvider router={router} />
+        <Toaster />
+        <UploadProgressWidget />
       </QueryClientProvider>
     </ThemeProvider>
   )
