+++ frontend/src/routes/_authenticated/knowledgeManagement.tsx
@@ -58,7 +58,7 @@ import {
 import ExcelViewer from "@/components/ExcelViewer"
 import CsvViewer from "@/components/CsvViewer"
 import TxtViewer from "@/components/TxtViewer"
-import { useUploadProgress } from "@/contexts/UploadProgressContext"
+import { useUploadProgress } from "@/store"
 import { DebugDocModal } from "@/components/DebugDocModal"
 
 // Persistent storage for documentId -> tempChatId mapping using sessionStorage
