+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "zustand": "^5.0.8"
+  }
+}
